+++ src/BankAccountNormalizer/SixDigitNormalizer.php
@@ -4,10 +4,11 @@ namespace Cs278\BankModulus\BankAccountNormalizer;
 
 use Cs278\BankModulus\BankAccountInterface;
 use Cs278\BankModulus\BankAccountNormalized;
+use Cs278\BankModulus\BankAccountNormalizedInterface;
 
 final class SixDigitNormalizer implements NormalizerInterface
 {
-    /** @return BankAccountInterface */
+    /** @return BankAccountInterface|BankAccountNormalizedInterface */
     public function normalize(BankAccountInterface $bankAccount)
     {
         return new BankAccountNormalized(