+++ CHANGELOG.md
@@ -9,10 +9,17 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Added
 - Implemented VocaLink V5.90 specification.
+- `BankAccountNormalized` now implements new interface `BankAccountNormalizedInterface`.
 
 ### Changed
 - Began testing on PHP 7.4.
 - Upgraded to PHPStan 0.12.x.
+
+### Deprecated
+- `NormalizerInterface::normalize()` will return `BankAccountNormalizedInterface`
+  in version 2.0.
+- `BankAccountNormalized::LENGTH` will be removed in version 2.0, replace with
+  `BankAccountNormalizedInterface::ACCOUNT_NUMBER_LENGTH`.
 
 ## [v1.10.0] - 2019-09-18